Final Refactoring
+ ranking dir contains now the new obsever implementation
- new_ranking removed
- previous ranking version is now archived localy

100% pytest passed
82% coverage
