--- rstt/new_ranking/__init__.py
@@ -1,2 +1,0 @@
-from .observer.game_observer import ObsTemplate, GameByGame, BatchGame
-from .standard import BTRanking, WinRate, BasicElo, BasicGlicko, BasicOS, SuccessRanking
