+++ rstt/ranking/inferer/__init__.py
@@ -1,0 +1,5 @@
+from .elo import Elo
+from .glicko import Glicko
+from .playerlevel import PlayerLevel
+from .playerwinprc import PlayerWinPRC
+from .eventscoring import EventScoring
