+++ rstt/ranking/inferer/playerlevel.py
@@ -1,0 +1,6 @@
+from rstt.stypes import SPlayer
+
+
+class PlayerLevel:
+    def rate(self, player: SPlayer) -> float:
+        return player.level()
