+++ rstt/ranking/observer/__init__.py
@@ -1,2 +1,3 @@
 from .obs import ObsTemplate
 from .game_observer import GameByGame, BatchGame
+from .player_observer import PlayerChecker, NoHandling
